refactor: mover arquivos para pasta risk
